APIs/googleapis.com/content/v2.1/python-aiohttp-client/openapi_server/controllers/liasettings_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server.models.lia_omnichannel_experience import LiaOmnichannelExperience
from openapi_server.models.lia_settings import LiaSettings
from openapi_server.models.liasettings_custom_batch_request import LiasettingsCustomBatchRequest
from openapi_server.models.liasettings_custom_batch_response import LiasettingsCustomBatchResponse
from openapi_server.models.liasettings_get_accessible_gmb_accounts_response import LiasettingsGetAccessibleGmbAccountsResponse
from openapi_server.models.liasettings_list_pos_data_providers_response import LiasettingsListPosDataProvidersResponse
from openapi_server.models.liasettings_list_response import LiasettingsListResponse
from openapi_server.models.liasettings_request_gmb_access_response import LiasettingsRequestGmbAccessResponse
from openapi_server.models.liasettings_request_inventory_verification_response import LiasettingsRequestInventoryVerificationResponse
from openapi_server.models.liasettings_set_inventory_verification_contact_response import LiasettingsSetInventoryVerificationContactResponse
from openapi_server.models.liasettings_set_pos_data_provider_response import LiasettingsSetPosDataProviderResponse
from openapi_server import util


async def content_liasettings_custombatch(request: web.Request, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, body=None) -> web.Response:
    """content_liasettings_custombatch

    Retrieves and/or updates the LIA settings of multiple accounts in a single request.

    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param body: 
    :type body: dict | bytes

    """
    body = LiasettingsCustomBatchRequest.from_dict(body)
    return web.Response(status=200)


async def content_liasettings_get(request: web.Request, merchant_id, account_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_get

    Retrieves the LIA settings of the account.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which to get or update LIA settings.
    :type account_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_getaccessiblegmbaccounts(request: web.Request, merchant_id, account_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_getaccessiblegmbaccounts

    Retrieves the list of accessible Business Profiles.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which to retrieve accessible Business Profiles.
    :type account_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_list(request: web.Request, merchant_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, max_results=None, page_token=None) -> web.Response:
    """content_liasettings_list

    Lists the LIA settings of the sub-accounts in your Merchant Center account.

    :param merchant_id: The ID of the managing account. This must be a multi-client account.
    :type merchant_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param max_results: The maximum number of LIA settings to return in the response, used for paging.
    :type max_results: int
    :param page_token: The token returned by the previous request.
    :type page_token: str

    """
    return web.Response(status=200)


async def content_liasettings_listposdataproviders(request: web.Request, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_listposdataproviders

    Retrieves the list of POS data providers that have active settings for the all eiligible countries.

    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_requestgmbaccess(request: web.Request, merchant_id, account_id, gmb_email, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_requestgmbaccess

    Requests access to a specified Business Profile.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which Business Profile access is requested.
    :type account_id: str
    :param gmb_email: The email of the Business Profile.
    :type gmb_email: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_requestinventoryverification(request: web.Request, merchant_id, account_id, country, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_requestinventoryverification

    Requests inventory validation for the specified country.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account that manages the order. This cannot be a multi-client account.
    :type account_id: str
    :param country: The country for which inventory validation is requested.
    :type country: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_setinventoryverificationcontact(request: web.Request, merchant_id, account_id, country, language, contact_name, contact_email, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None) -> web.Response:
    """content_liasettings_setinventoryverificationcontact

    Sets the inventory verification contract for the specified country.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account that manages the order. This cannot be a multi-client account.
    :type account_id: str
    :param country: The country for which inventory verification is requested.
    :type country: str
    :param language: The language for which inventory verification is requested.
    :type language: str
    :param contact_name: The name of the inventory verification contact.
    :type contact_name: str
    :param contact_email: The email of the inventory verification contact.
    :type contact_email: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str

    """
    return web.Response(status=200)


async def content_liasettings_setomnichannelexperience(request: web.Request, merchant_id, account_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, country=None, lsf_type=None, pickup_types=None) -> web.Response:
    """content_liasettings_setomnichannelexperience

    Sets the omnichannel experience for the specified country. Only supported for merchants whose POS data provider is trusted to enable the corresponding experience. For more context, see these help articles [about LFP](https://support.google.com/merchants/answer/7676652) and [how to get started](https://support.google.com/merchants/answer/7676578) with it.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which to retrieve accessible Business Profiles.
    :type account_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param country: The CLDR country code (for example, \&quot;US\&quot;) for which the omnichannel experience is selected.
    :type country: str
    :param lsf_type: The Local Store Front (LSF) type for this country. Acceptable values are: - \&quot;&#x60;ghlsf&#x60;\&quot; (Google-Hosted Local Store Front) - \&quot;&#x60;mhlsfBasic&#x60;\&quot; (Merchant-Hosted Local Store Front Basic) - \&quot;&#x60;mhlsfFull&#x60;\&quot; (Merchant-Hosted Local Store Front Full) More details about these types can be found here.
    :type lsf_type: str
    :param pickup_types: The Pickup types for this country. Acceptable values are: - \&quot;&#x60;pickupToday&#x60;\&quot; - \&quot;&#x60;pickupLater&#x60;\&quot; 
    :type pickup_types: List[str]

    """
    return web.Response(status=200)


async def content_liasettings_setposdataprovider(request: web.Request, merchant_id, account_id, country, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, pos_data_provider_id=None, pos_external_account_id=None) -> web.Response:
    """content_liasettings_setposdataprovider

    Sets the POS data provider for the specified country.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which to retrieve accessible Business Profiles.
    :type account_id: str
    :param country: The country for which the POS data provider is selected.
    :type country: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param pos_data_provider_id: The ID of POS data provider.
    :type pos_data_provider_id: str
    :param pos_external_account_id: The account ID by which this merchant is known to the POS data provider.
    :type pos_external_account_id: str

    """
    return web.Response(status=200)


async def content_liasettings_update(request: web.Request, merchant_id, account_id, xgafv=None, access_token=None, alt=None, param_callback=None, fields=None, key=None, oauth_token=None, pretty_print=None, quota_user=None, upload_protocol=None, upload_type=None, body=None) -> web.Response:
    """content_liasettings_update

    Updates the LIA settings of the account. Any fields that are not provided are deleted from the resource.

    :param merchant_id: The ID of the managing account. If this parameter is not the same as accountId, then this account must be a multi-client account and &#x60;accountId&#x60; must be the ID of a sub-account of this account.
    :type merchant_id: str
    :param account_id: The ID of the account for which to get or update LIA settings.
    :type account_id: str
    :param xgafv: V1 error format.
    :type xgafv: str
    :param access_token: OAuth access token.
    :type access_token: str
    :param alt: Data format for response.
    :type alt: str
    :param param_callback: JSONP
    :type param_callback: str
    :param fields: Selector specifying which fields to include in a partial response.
    :type fields: str
    :param key: API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.
    :type key: str
    :param oauth_token: OAuth 2.0 token for the current user.
    :type oauth_token: str
    :param pretty_print: Returns response with indentations and line breaks.
    :type pretty_print: bool
    :param quota_user: Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.
    :type quota_user: str
    :param upload_protocol: Upload protocol for media (e.g. \&quot;raw\&quot;, \&quot;multipart\&quot;).
    :type upload_protocol: str
    :param upload_type: Legacy upload protocol for media (e.g. \&quot;media\&quot;, \&quot;multipart\&quot;).
    :type upload_type: str
    :param body: 
    :type body: dict | bytes

    """
    body = LiaSettings.from_dict(body)
    return web.Response(status=200)
